requirements.txt
Django>=3.0.6
psycopg2>=2.8.5
django-ckeditor>=5.9.0
